// packages/protobuf-test/src/gen/ts/extra/wkt-any_pb.ts
// Copyright 2021-2022 Buf Technologies, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-es v0.2.0 with parameter "ts_nocheck=false,target=ts"
// @generated from file extra/wkt-any.proto (package spec, syntax proto3)
/* eslint-disable */

import type {BinaryReadOptions, FieldList, JsonReadOptions, JsonValue, PartialMessage, PlainMessage} from "@bufbuild/protobuf";
import {Any, Message, proto3} from "@bufbuild/protobuf";

/**
 * @generated from message spec.MessageWithAnyField
 */
export class MessageWithAnyField extends Message<MessageWithAnyField> {
  /**
   * @generated from field: google.protobuf.Any field = 1;
   */
  field?: Any;

  constructor(data?: PartialMessage<MessageWithAnyField>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime = proto3;
  static readonly typeName = "spec.MessageWithAnyField";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "field", kind: "message", T: Any },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MessageWithAnyField {
    return new MessageWithAnyField().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MessageWithAnyField {
    return new MessageWithAnyField().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MessageWithAnyField {
    return new MessageWithAnyField().fromJsonString(jsonString, options);
  }

  static equals(a: MessageWithAnyField | PlainMessage<MessageWithAnyField> | undefined, b: MessageWithAnyField | PlainMessage<MessageWithAnyField> | undefined): boolean {
    return proto3.util.equals(MessageWithAnyField, a, b);
  }
}